Changed type of FrontendLabel store_id to Integer

+++ src/main/java/com/github/chen0040/magento/models/product/FrontendLabel.java
@@ -11,5 +11,5 @@ import lombok.Setter;
 @AllArgsConstructor
 public class FrontendLabel {
 	String label;
-	Long store_id;
+	Integer store_id;
 }
